requirements.txt
Pillow==9.0.1
pre-commit==2.17.0
